Completed coding the VIIRS path but receiving error when processing the masking. Error about extents not overlapping.
Also requires to have polygon download put in the right place.

Also requires a more efficient path. Where space is not an issue pre-download all tiles and polygons and process as necessary. Where long-term space is an issue but short-term is okay download all, process and download since nightlight data does not change then allow individual nightlight processes to be done on the fly.